factor container config out of container creation, refactor unit tests
